useWindowScroll hook, update sticky elements offset
