+++ src/components/layouts/SuperHeader/SuperHeader.js
@@ -1,7 +1,9 @@
-import React, { useState } from 'react'
+import React from 'react'
 
 import { Box } from '@material-ui/core'
 import { withStyles, createStyles } from '@material-ui/core/styles'
+
+import useWindowScroll from '../../../js/hooks/useWindowScroll'
 
 import InfoBanner from '../../content-partials/InfoBanner'
 import ShutdownBanner from '../../content-partials/ShutdownBanner'
@@ -18,10 +20,10 @@ const SuperHeaderContainer = withStyles(theme =>
       width: '100vw',
       zIndex: '1002',
       maxWidth: '100%',
-      transition: 'height .2s ease',
+      transition: 'height .1s ease',
       '& .notActive > div:first-child': {
         height: 0,
-        transition: 'height .1s ease',
+        transition: 'height .2s ease',
       },
       '& .notActive > div:last-child > header > div': {
         height: 60,
@@ -38,41 +40,10 @@ const SuperHeaderContainer = withStyles(theme =>
 )(Box)
 
 const SuperHeader = ({ data, ...rest }) => {
-  const [isActive, setIsActive] = useState(true)
-
-  React.useEffect(() => {
-    let scrollPosition = 0
-
-    const pageHeight = document.body.offsetHeight
-    const viewportHeight = window.innerHeight
-
-    function handleScroll () {
-      const newScrollPosition = window.scrollY
-
-      if (newScrollPosition === scrollPosition) {
-        return
-      }
-
-      if (newScrollPosition < 0 || newScrollPosition + viewportHeight > pageHeight) {
-        return
-      }
-
-      const shouldShow = newScrollPosition < scrollPosition
-      setIsActive(shouldShow)
-
-      scrollPosition = newScrollPosition
-    }
-
-    window.addEventListener('scroll', handleScroll, { passive: true })
-
-    return () => {
-      window.removeEventListener('scroll', handleScroll)
-    }
-  }, [])
-
+  const isScrollActive = useWindowScroll()
   return (
     <SuperHeaderContainer>
-      <div className={isActive ? 'isActive' : 'notActive'}>
+      <div className={isScrollActive ? 'isActive' : 'notActive'}>
         <InfoBanner />
         {data.site.siteMetadata.isShutdown === 'true' &&
           <ShutdownBanner />
