+++ src/components/navigation/PageSubMenu/PageSubMenu.js
@@ -144,7 +144,7 @@ const PageSubMenu = ({ children, menuItems, ...props }) => {
   return (
     <>
       <Box className={classes.root}>
-        <StickyWrapper enabled={true} top={120} bottomBoundary={0} innerZ="1000" activeClass="sticky">
+        <StickyWrapper enabled={true} top={180} bottomBoundary={0} innerZ="1000" activeClass="sticky">
           <Paper elevation={1} square>
             <Container maxWidth="lg">
               <MenuList id="page-scrollto-subnav">
